# [engraving] improved layout performance