--- a/src/engraving/infrastructure/ld_access.h
+++ b/src/engraving/infrastructure/ld_access.h
@@ -51,18 +51,18 @@ enum class LD_ACCESS {
 #define LD_INDEPENDENT
 
 template<typename T>
-class ld_field
+class ld_field_debug
 {
 public:
 
-    ld_field(const char* name, T def = T())
+    ld_field_debug(const char* name, T def = T())
         : m_name(name), m_def(def) {}
 
-    void reset() { m_val.reset(); }
+    inline void reset() { m_val.reset(); }
 
-    bool has_value() const { return m_val.has_value(); }
+    inline bool has_value() const { return m_val.has_value(); }
 
-    const T& value(LD_ACCESS mode = LD_ACCESS::CHECK) const
+    inline const T& value(LD_ACCESS mode = LD_ACCESS::CHECK) const
     {
         if (!m_val.has_value()) {
 #ifdef MUE_ENABLE_ENGRAVING_LD_ACCESS
@@ -77,12 +77,12 @@ public:
         return m_val.value();
     }
 
-    const T& operator()() const
+    inline const T& operator()() const
     {
         return value();
     }
 
-    T& mut_value()
+    inline T& mut_value()
     {
         if (!m_val.has_value()) {
             m_val = std::make_optional<T>(m_def);
@@ -90,7 +90,7 @@ public:
         return m_val.value();
     }
 
-    void set_value(const T& v)
+    inline void set_value(const T& v)
     {
         m_val = std::make_optional<T>(v);
     }
@@ -100,6 +100,49 @@ private:
     T m_def;
     std::optional<T> m_val;
 };
+
+template<typename T>
+class ld_field_prod
+{
+public:
+
+    ld_field_prod(const char*, T = T()) {}
+
+    inline void reset() { m_val = T(); }
+
+    inline bool has_value() const { return true; }
+
+    inline const T& value(LD_ACCESS = LD_ACCESS::CHECK) const
+    {
+        return m_val;
+    }
+
+    inline const T& operator()() const
+    {
+        return value();
+    }
+
+    inline T& mut_value()
+    {
+        return m_val;
+    }
+
+    inline void set_value(const T& v)
+    {
+        m_val = v;
+    }
+
+private:
+    T m_val;
+};
+
+#ifdef NDEBUG
+template<typename T>
+using ld_field = ld_field_prod<T>;
+#else
+template<typename T>
+using ld_field = ld_field_debug<T>;
+#endif
 }
 
 #endif // MU_ENGRAVING_LD_ACCESS_DEV_H