Autobot TC9: Big Score (perfomance)

4.1.1
```
Function                                                    Call time         Call count        Sum time          
ScriptEngine::doCall                                        13006.062 ms      1                 13006.062 ms      
ShortcutsApi::activate                                      4384.707 ms       2                 8769.413 ms       
NotationActionController::addSlur                           7713.050 ms       1                 7713.050 ms       
NotationInteraction::notifyAboutSelectionChangedIfNeed      600.167 ms        3                 1800.500 ms       
InspectorListModel::setElementList                          347.597 ms        5                 1737.984 ms       
AutobotApi::openProject                                     1683.117 ms       1                 1683.117 ms       
ProjectActionsController::doOpenProject                     1208.032 ms       1                 1208.032 ms       
Score::update                                               258.909 ms        4                 1035.636 ms       
Score::doLayoutRange                                        505.626 ms        2                 1011.252 ms     
ProjectActionsController::loadProject                       1000.873 ms       1                 1000.873 ms       
NotationProject::load                                       1000.829 ms       1                 1000.829 ms       
NotationProject::doLoad                                     1000.512 ms       1                 1000.512 ms       
SystemLayout::collectSystem                                 17.394 ms         56                974.045 ms        
PageLayout::collectPage                                     17.676 ms         54                954.528 ms  
```

4.2 (this PR)
```
Function                                                    Call time         Call count        Sum time          
ScriptEngine::doCall                                        9073.113 ms       1                 9073.113 ms       
ShortcutsApi::activate                                      2392.610 ms       2                 4785.220 ms       
NotationActionController::addSlur                           4737.388 ms       1                 4737.388 ms       
AutobotApi::openProject                                     1753.390 ms       1                 1753.390 ms       
ProjectActionsController::doOpenProject                     1277.553 ms       1                 1277.553 ms       
Score::update                                               289.903 ms        4                 1159.613 ms       
Score::doLayoutRange                                        568.719 ms        2                 1137.439 ms       
ScoreLayout::layoutRange                                    567.845 ms        2                 1135.691 ms       
SystemLayout::collectSystem                                 19.635 ms         56                1099.534 ms       
ProjectActionsController::loadProject                       1089.886 ms       1                 1089.886 ms       
NotationProject::load                                       1089.844 ms       1                 1089.844 ms       
NotationProject::doLoad                                     1089.449 ms       1                 1089.449 ms       
PageLayout::collectPage                                     19.852 ms         54                1071.988 ms  
```

Layout run a little slower, but within measurement volatility.

Measurement volatility is quite high :( 
So I'm planning improve a performance test (separate PR)