@mike-spa 2023-11-22
I've mentioned this before, but I would suggest not using that test for measuring layout performance. What that test does is open a big file and add a huge number of slurs to it. Indeed you can see that the biggest time component is `NotationActionController::addSlur`, which is not really what we are interested in (and is also highly unrealistic as a use case). We are only interested in `Score::doLayoutRange`, that is where all the layout is done.

The best performance test we can do for layout is to simply open a very big score with real-world notation. Beethoven 9 is probably the best score we have for this [Beethoven_SymphonyN9_MU4.zip](https://github.com/musescore/MuseScore/files/13440281/Beethoven_SymphonyN9_MU4.zip). There you will see that the n.1 time consuming process is doLayoutRange, then we'll have a more realistic measurement :)