@igorkorsukov 2023-11-23
> Why those vtest differences?

fixed